Points can now be undone
Delete latest point deletes the point added latest to the bezier curve added
latest. If it is the ending point of a curve also both control points are
deleted. Makes no sense to have control points for a non existent curve.
If it is the starting point of a curve the whole curve is deleted. No one
needs curves without points. If there are no curves, meaning there are no
points nothing happens. If the ending point of a curve is removed the next
point added will become the new ending point of the curve.