Move points with rightclick on them and holding right mouse button down.
New points are still added with left click but to move them click right onto
them and hold the right mousebutton down. This now allows to place different
points at the same location. The later added points are on top of the
previously added points. Also enlarged select radius for points so it is
easiert to drag them around.