Right click on existing start or end point => use its coords for new point
When user rightclicks on existing start or end point of another curve the
exact coords of this point are used for the new point. When user later left
clicks on these points to drag them, they move around as one point. But when
removing the last point they are still removed as two points.
For now there is this little bug when rightclicking on an existing control
point the new point and this control point later also move as one point. But
I will introduce a feature to select and deselect curves to show and hide
their contorl points later on. This should then also fix this little bug. If not
I will try to find another solution.