Added possibility to scale the canvas and the drawings on it.
Initially scaling is 1 but scaling can be changed anytime and then the canvas
is resized and the curves are redrawn accordingly. The inital size of the
canvas is always the target size. So even something is drawn on the bigger
canvas internally all curves are stored as if they were drawn on the inital
size of the canvas and then rescaled for drawing on the bigger canvas.
This makes it possible to change the scaling as often as one wants to.